Create first draft, squares colored on hover
A 4x4 grid of 'div's are created by the JS function. On hover, they turn
black.

With a horizontal aspect ratio for the browser, this will result in an
unused white space on the right side of the browser.
With a vertical aspect ratio, the entire row of squares will not be
displayed.

TODO: Center the whole thing.